api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pyyaml==6.0.1
numpy==1.26.4
torch==2.2.0
pydantic==2.7.1